Refactor: 사용하지 않는 파일 제거
사용하지 않는 controller, service 제거
사용하지 않는 함수 및 호출구문 제거
