--- src/business-hour/business-hour.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('business-hour')
-export class BusinessHourController {}
