--- src/business-hour/business-hour.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BusinessHourController } from './business-hour.controller';
-import { BusinessHourService } from './business-hour.service';
-
-@Module({
-  controllers: [BusinessHourController],
-  providers: [BusinessHourService],
-})
-export class BusinessHourModule {}
