--- src/business-hour/business-hour.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class BusinessHourService {}
