--- src/reservation-events/dto/accept-seat.dto.ts
@@ -1,5 +1,0 @@
-export class AcceptSeatDTO {
-  userId: string;
-
-  reservationId: number;
-}
