--- src/store-table/store-table.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('store-table')
-export class StoreTableController {}
