--- src/store-table/store-table.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { StoreTableController } from './store-table.controller';
-import { StoreTableService } from './store-table.service';
-
-@Module({
-  controllers: [StoreTableController],
-  providers: [StoreTableService],
-})
-export class StoreTableModule {}
